requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
xlrd>=2.0.0
xlutils>=2.0.0
xlwt>=1.3.0
openpyxl>=3.1.0

# Postgres
psycopg[binary]>=3.1
psycopg_pool>=3.2

